pyproject.toml
[project]
name = "aline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "feedparser>=6.0.11",
    "greenlet>=3.2.1",
    "line-bot-sdk>=3.17.1",
    "litellm>=1.67.2",
    "motor>=3.7.0",
    "openai-agents>=0.0.13",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
    "weave[scorers]>=0.51.44",
]
